+++ gserver/src/steps.handler.ts
@@ -206,8 +206,8 @@ export default class StepsHandler {
         //https://docs.cucumber.io/cucumber/cucumber-expressions/
         step = step.replace(/{float}/g, '-?\\d*\\.?\\d+');
         step = step.replace(/{int}/g, '-?\\d+');
-        step = step.replace(/{stringInDoubleQuotes}/g, '"[^"]+"');
-        step = step.replace(/{string}/g, '"[^"]+"');
+        step = step.replace(/{word}/g, '[A-Za-z]+');
+        step = step.replace(/{string}/g, '[\"|\'].*[\"|\']');
         step = step.replace(/{}/g, '.*');
 
         //Optional Text