+++ gserver/test/data/steps/cucumberExpressions.steps.js
@@ -1,0 +1,33 @@
+//https://github.com/cucumber/cucumber-expressions
+
+this.When("I have a {int} in my belly", function (next) {
+    next;
+});
+
+this.When("I have a {float} in my belly", function (next) {
+    next;
+});
+
+this.When("I have a important {word} in my belly", function (next) {
+    next;
+});
+
+this.When("I have a {string} in my belly", function (next) {
+    next;
+});
+
+this.When("I have a wildcard {} in my belly", function (next) {
+    next;
+});
+
+this.When("I have a cucumber in my belly/stomach", function (next) {
+    next;
+});
+
+this.When("I have a gherkin(s) in my belly", function (next) {
+    next;
+});
+
+this.When("I have a cucumber\(s\) in my belly", function (next) {
+    next;
+});